server/requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Redis + background jobs
redis>=5.2.0
arq>=0.26.1

# Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Payments (Phase 7)
# stripe>=11.0.0

# Dev
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
